[mr2423] rewrite MapObjectsStore v0.13   - use GuidSet for make object-safe UpdateDataMapType (this fix at least pet action bar after mount)   - use unordered_map instead of TypeUnorderedMapContainer template collection   - full rewrite object update making system, make dependent update queue for players items   - add Transport objects to current Map storing/updating system (currently by half-hack)   - simplify code   - use storing guids instead of pointers in Camera/Viewpoint subsystem for fit new object storing system   - fix very old bug with Sentry totem (appears only now).
    NEED TEST and measurements CPU load and resource leaks! related to #1165, #1083, #852

+++ src/game/Unit.cpp
@@ -288,6 +288,9 @@ Unit::Unit() :
 
 Unit::~Unit()
 {
+    if (IsInWorld())
+        WorldObject::RemoveFromWorld(true);
+
     // set current spells as deletable
     for (uint32 i = 0; i < CURRENT_MAX_SPELL; ++i)
     {
@@ -11197,11 +11200,11 @@ uint32 Unit::GetCreatePowers(Powers power) const
 
 void Unit::AddToWorld()
 {
-    Object::AddToWorld();
+    WorldObject::AddToWorld();
     ScheduleAINotify(0);
 }
 
-void Unit::RemoveFromWorld()
+void Unit::RemoveFromWorld(bool remove)
 {
     // cleanup
     if (IsInWorld())
@@ -11213,22 +11216,26 @@ void Unit::RemoveFromWorld()
         UnsummonAllTotems();
         RemoveAllGameObjects();
         RemoveAllDynObjects();
-        CleanupDeletedAuras();
-        GetViewPoint().Event_RemovedFromWorld();
+        CleanupDeletedAuras();        
     }
+    GetViewPoint().Event_RemovedFromWorld();
 
-    Object::RemoveFromWorld();
+    WorldObject::RemoveFromWorld(remove);
 }
 
 void Unit::CleanupsBeforeDelete()
 {
+    if (!IsInWorld())
+        return;
+
     if (m_uint32Values)                                     // only for fully created object
     {
         InterruptNonMeleeSpells(true);
         m_Events.KillAllEvents(false);                      // non-delatable (currently casted spells) will not deleted now but it will deleted at call in Map::RemoveAllObjectsInRemoveList
         CombatStop();
         ClearComboPointHolders();
-        DeleteThreatList();
+        if (CanHaveThreatList())
+            DeleteThreatList();
         if (GetTypeId() == TYPEID_PLAYER)
             getHostileRefManager().setOnlineOfflineState(false);
         else
